requirements.txt
apturl==0.5.2
beautifulsoup4==4.4.1
blinker==1.3
Brlapi==0.6.4
certifi==2018.11.29
chardet==3.0.4
checkbox-support==0.22
clearbit==0.1.7
command-not-found==0.3
cryptography==1.2.3
defer==1.0.6
Django==2.1.5
django-filter==2.1.0
django-rest-auth==0.9.3
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.2.1
feedparser==5.1.3
guacamole==0.9.2
gunicorn==19.9.0
html5lib==0.999
httplib2==0.9.1
idna==2.8
Jinja2==2.8
language-selector==0.1
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==3.0.1
MarkupSafe==0.23
oauthlib==1.0.3
onboard==1.2.0
padme==1.1.1
pexpect==4.0.1
Pillow==3.1.2
plainbox==0.25
psycopg2==2.7.3.2
ptyprocess==0.5
public==2018.11.20
pyasn1==0.1.9
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
pyhunter==0.3
PyJWT==1.5.3
pyparsing==2.0.3
python-apt==1.1.0b1+ubuntu0.16.4.2
python-debian==0.1.27
python-systemd==231
pytz==2018.9
pyxdg==0.25
reportlab==3.3.0
requests==2.20.0
screen-resolution-extra==0.0.0
sessioninstaller==0.0.0
six==1.11.0
system-service==0.3
timedelta==2018.11.20
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
usb-creator==0.3.0
username-generator==2.0.0
virtualenv==15.0.1
xdiagnose==3.8.4.1
XlsxWriter==0.7.3

